third commiit

+++ server/src/main/java/com/cognizant/pm/repository/ParentTaskRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.cognizant.pm.dao.ParentTask;
 
-public interface ParentTaskRepository extends CrudRepository<ParentTask, Long> {
+public interface ParentTaskRepository extends CrudRepository<ParentTask, Integer> {
 
 }
